+++ src/main/java/nextstep/entity/Theme.java
@@ -1,9 +1,12 @@
-package nextstep.domain;
+package nextstep.entity;
 
 public class Theme {
-    private String name;
-    private String desc;
-    private Integer price;
+
+    private final String name;
+
+    private final String desc;
+
+    private final Integer price;
 
     public Theme(String name, String desc, Integer price) {
         this.name = name;
